requirements.txt
beautifulsoup4
Flask
Flask-Assets
GitPython
markdown-it-py
mdit-py-plugins
osu-wiki-tools
PyYAML
